+++ app/Http/Controllers/Auth/RegisterController.php
@@ -67,8 +67,6 @@ class RegisterController extends Controller
             'name' => $data['name'],
             'email' => $data['email'],
             'password' => Hash::make($data['password']),
-        ] + [
-            'is_horeca' => request()->has('is_horeca')
         ]);
     }
 }