+++ CHANGELOG.md
@@ -10,7 +10,8 @@ Log of unreleased changes.
 
 - Adds `isHoreca()` method on `User` model.
 - Removes `is_horeca` field on `User` model.
-- Adds `legal_person` filter on `BillingProfile` model. 
+- Adds `legal_person` filter on `BillingProfile` model.
+- Removes `choice.blade.php` view. 
 
 ### Fixed